requirements.txt
pony==0.7.14
python-decouple
passlib[bcrypt]
pyjwt
redis
#mysqlclient==2.1.1
PyMySQL==1.0.2
fastapi-mail==0.4.1
python-dotenv==0.19.0
#google-api-python-client
gunicorn==20.0.4
uvicorn==0.15.0
fastapi==0.68.1
python-multipart==0.0.5
python-dateutil==2.8.2
#pycryptodome
idna==2.8
h11==0.12
protobuf==3.20.1
boto3==1.21.38
#nltk==3.7
fastapi-versioning==0.10.0
# firebase-admin==5.2.0